Only regenerate API resource pages if modified
Only generate API resource pages if files in the RAML directory for a
given API version, in the "apps" directory or in "_layouts" have been
modified.

With this changes, if only a blog post is modified or added, it will
regenerate approximately 15 seconds faster (takes circa 3 seconds with
Jekyll's `--incremental` option), because previously all API resource
pages were regenerated with every change even if no RAML files were
modified.

This might yield "Unknown reference: Could not find page with key"
warnings in dev mode (`rake dev`) in the case a blog post referencing
an API resource page is modified, but works for common blog edits
(usually not linking to an API resource page).